Add MT19937 random generator. devfs: Add `random`, `urandom` device. page: `PageManagement::set_pte_addr` -> `PageManagement::map`, and add `umap` function.

+++ src/arch/riscv64/cpu.rs
@@ -8,74 +8,74 @@ use asm_wrap::*;
 pub mod asm_wrap {
     use core::arch::asm;
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn mie_r() -> u64 {
         let mut mie: u64;
         asm!("csrr {}, mie", out(reg) mie);
         mie
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn mie_w(mie: u64) {
         asm!("csrw mie, {}", in(reg) mie);
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn sie_r() -> u64 {
         let mut sie: u64;
         asm!("csrr {}, sie", out(reg) sie);
         sie
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn sie_w(sie: u64) {
         asm!("csrw sie, {}", in(reg) sie);
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn sstatus_r() -> u64 {
         let mut sstatus: u64;
         asm!("csrr {}, sstatus", out(reg) sstatus);
         sstatus
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn sstatus_w(sstatus: u64) {
         asm!("csrw sstatus, {}", in(reg) sstatus);
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn mepc_r() -> u64 {
         let mut mepc: u64;
         asm!("csrr {}, mepc", out(reg) mepc);
         mepc
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn mepc_w(mepc: u64) {
         asm!("csrw mepc, {}", in(reg) mepc);
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn sepc_r() -> u64 {
         let mut sepc: u64;
         asm!("csrr {}, sepc", out(reg) sepc);
         sepc
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn sepc_w(sepc: u64) {
         asm!("csrw sepc, {}", in(reg) sepc);
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn mtvec_r() -> u64 {
         let mut mtvec: u64;
         asm!("csrr {}, mtvec", out(reg) mtvec);
         mtvec
     }
 
-    #[inline]
+    #[inline(always)]
     pub unsafe fn mtvec_w(mtvec: u64) {
         asm!("csrw mtvec, {}", in(reg) mtvec);
     }
@@ -104,7 +104,7 @@ unsafe fn trap_switch_to_s_level() {
 
 pub unsafe fn switch_to_s_level() {
     let mtvec = mtvec_r();
-    mtvec_w(trap_switch_to_s_level as u64);
+    mtvec_w(trap_switch_to_s_level as usize as u64);
     asm!("ecall");
     mtvec_w(mtvec);
 }
